+++ lab1.h
@@ -9,11 +9,11 @@
 namespace SundayWork {
 void eulerMethod(std::fstream& fileLab1,
                  double startX, double startY, double endX, double step,
-                 std::function<long double(double, double)> func);
+                 std::function<long double(long double, long double)> func);
 
 void modifiedEulerMethod(std::fstream& fileLab1,
                          double startX, double startY, double endX, double step,
-                         std::function<long double(double, double)> func);
+                         std::function<long double(long double, long double)> func);
 }
 
 #endif // LAB1_H
